Added forkscreen if no households
